+++ examples/data-sources/harness_platform_file_store_file/data-source.tf
@@ -1,0 +1,3 @@
+data "harness_platform_file_store_file" "example" {
+  identifier = "identifier"
+}