+++ examples/resources/harness_platform_file_store_file/import.sh
@@ -1,0 +1,8 @@
+# Import account level file
+terraform import harness_platform_file_store_file.example <identifier>
+
+# Import org level file
+terraform import harness_platform_file_store_file.example <org_id><identifier>
+
+# Import org level file
+terraform import harness_platform_file_store_file.example <org_id><project_id><identifier>