+++ internal/service/platform/template/resource_template.go
@@ -27,7 +27,8 @@ func ResourceTemplate() *schema.Resource {
 			"template_yaml": {
 				Description: "Yaml for creating new Template." + helpers.Descriptions.YamlText.String(),
 				Type:        schema.TypeString,
-				Required:    true,
+				Optional: true,
+				Computed: true,
 			},
 			"version": {
 				Description: "Version Label for Template.",
@@ -38,6 +39,7 @@ func ResourceTemplate() *schema.Resource {
 				Description: "True if given version for template to be set as stable.",
 				Type:        schema.TypeBool,
 				Optional:    true,
+				Computed: true,
 			},
 			"comments": {
 				Description: "Specify comment with respect to changes.",
@@ -150,6 +152,71 @@ func ResourceTemplate() *schema.Resource {
 				},
 				Optional: true,
 			},
+			"import_from_git": {
+				Description: "Flag to set if importing from Git",
+				Type:        schema.TypeBool,
+				Optional:    true,
+			},
+			"git_import_details": {
+				Description: "Contains Git Information for importing entities from Git",
+				Type:        schema.TypeList,
+				MaxItems:    1,
+				Optional:    true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"branch_name": {
+							Description: "Name of the branch.",
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+						"file_path": {
+							Description: "File path of the Entity in the repository.",
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+						"connector_ref": {
+							Description: "Identifier of the Harness Connector used for importing entity from Git" + helpers.Descriptions.ConnectorRefText.String(),
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+						"repo_name": {
+							Description: "Name of the repository.",
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+						"is_force_import": {
+							Description: "",
+							Type:        schema.TypeBool,
+							Optional:    true,
+						},
+					},
+				},
+			},
+			"template_import_request": {
+				Description: "Contains parameters for importing template.",
+				Type:        schema.TypeList,
+				MaxItems:    1,
+				Optional:    true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"template_name": {
+							Description: "Name of the template.",
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+						"template_version": {
+							Description: "Version of the template.",
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+						"template_description": {
+							Description: "Description of the template.",
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+					},
+				},
+			},
 		},
 	}
 
@@ -243,45 +310,73 @@ func resourceTemplateCreateOrUpdate(ctx context.Context, d *schema.ResourceData,
 	project_id := d.Get("project_id").(string)
 	comments := d.Get("comments").(string)
 	version := d.Get("version").(string)
+	template_yaml := d.Get("template_yaml").(string)
+	is_stable := d.Get("is_stable").(bool)
 
 	if id == "" {
-		template := buildCreateTemplate(d)
-		if template.GitDetails != nil {
-			base_branch = optional.NewString(template.GitDetails.BaseBranch)
-			store_type = optional.NewString(template.GitDetails.StoreType)
-			commit_message = optional.NewString(template.GitDetails.CommitMessage)
-			connector_ref = optional.NewString(template.GitDetails.ConnectorRef)
-			branch_name = template.GitDetails.BranchName
-		}
-		if project_id != "" {
-			resp, httpResp, err = c.ProjectTemplateApi.CreateTemplatesProject(ctx, org_id, project_id, &nextgen.ProjectTemplateApiCreateTemplatesProjectOpts{
-				Body:           optional.NewInterface(template),
-				HarnessAccount: optional.NewString(c.AccountId),
-			})
-			if resp.Identifier != "" {
-				template_id = resp.Identifier
+
+		if d.Get("import_from_git").(bool) {
+
+			template_id = d.Get("identifier").(string)
+
+			template_import_request_body := createImportFromGitRequestForTemplates(d)
+
+			if project_id != "" {
+				_, httpResp, err = c.ProjectTemplateApi.ImportProjectTemplatesFromGit(ctx, org_id, project_id, template_id,
+					&nextgen.TemplatesApiImportProjectTemplatesFromGitOpts{
+						Body:           optional.NewInterface(template_import_request_body),
+						HarnessAccount: optional.NewString(c.AccountId)})
+			} else if org_id != "" && project_id == "" {
+				_, httpResp, err = c.OrgTemplateApi.ImportOrgTemplatesFromGit(ctx, org_id, template_id,
+					&nextgen.TemplatesApiImportOrgTemplatesFromGitOpts{
+						Body:           optional.NewInterface(template_import_request_body),
+						HarnessAccount: optional.NewString(c.AccountId)})
 			} else {
-				template_id = resp.Slug
+				_, httpResp, err = c.AccountTemplateApi.ImportAccountLevelTemplatesFromGit(ctx, template_id,
+					&nextgen.TemplatesApiImportTemplateFromGitOpts{
+						Body:           optional.NewInterface(template_import_request_body),
+						HarnessAccount: optional.NewString(c.AccountId)})
 			}
-		} else if org_id != "" && project_id == "" {
-			resp, httpResp, err = c.OrgTemplateApi.CreateTemplatesOrg(ctx, org_id, &nextgen.OrgTemplateApiCreateTemplatesOrgOpts{
-				HarnessAccount: optional.NewString(c.AccountId),
-				Body:           optional.NewInterface(template),
-			})
-			if resp.Identifier != "" {
-				template_id = resp.Identifier
-			} else {
-				template_id = resp.Slug
+
+		} else{
+			template := buildCreateTemplate(d)
+			if template.GitDetails != nil {
+				base_branch = optional.NewString(template.GitDetails.BaseBranch)
+				store_type = optional.NewString(template.GitDetails.StoreType)
+				commit_message = optional.NewString(template.GitDetails.CommitMessage)
+				connector_ref = optional.NewString(template.GitDetails.ConnectorRef)
+				branch_name = template.GitDetails.BranchName
 			}
-		} else {
-			resp, httpResp, err = c.AccountTemplateApi.CreateTemplatesAcc(ctx, &nextgen.AccountTemplateApiCreateTemplatesAccOpts{
-				HarnessAccount: optional.NewString(c.AccountId),
-				Body:           optional.NewInterface(template),
-			})
-			if resp.Identifier != "" {
-				template_id = resp.Identifier
+			if project_id != "" {
+				resp, httpResp, err = c.ProjectTemplateApi.CreateTemplatesProject(ctx, org_id, project_id, &nextgen.ProjectTemplateApiCreateTemplatesProjectOpts{
+					Body:           optional.NewInterface(template),
+					HarnessAccount: optional.NewString(c.AccountId),
+				})
+				if resp.Identifier != "" {
+					template_id = resp.Identifier
+				} else {
+					template_id = resp.Slug
+				}
+			} else if org_id != "" && project_id == "" {
+				resp, httpResp, err = c.OrgTemplateApi.CreateTemplatesOrg(ctx, org_id, &nextgen.OrgTemplateApiCreateTemplatesOrgOpts{
+					HarnessAccount: optional.NewString(c.AccountId),
+					Body:           optional.NewInterface(template),
+				})
+				if resp.Identifier != "" {
+					template_id = resp.Identifier
+				} else {
+					template_id = resp.Slug
+				}
 			} else {
-				template_id = resp.Slug
+				resp, httpResp, err = c.AccountTemplateApi.CreateTemplatesAcc(ctx, &nextgen.AccountTemplateApiCreateTemplatesAccOpts{
+					HarnessAccount: optional.NewString(c.AccountId),
+					Body:           optional.NewInterface(template),
+				})
+				if resp.Identifier != "" {
+					template_id = resp.Identifier
+				} else {
+					template_id = resp.Slug
+				}
 			}
 		}
 	} else {
@@ -293,36 +388,59 @@ func resourceTemplateCreateOrUpdate(ctx context.Context, d *schema.ResourceData,
 			commit_message = optional.NewString(template.GitDetails.CommitMessage)
 			connector_ref = optional.NewString(template.GitDetails.ConnectorRef)
 		}
-		if project_id != "" {
-			resp, httpResp, err = c.ProjectTemplateApi.UpdateTemplateProject(ctx, project_id, id, org_id, version, &nextgen.ProjectTemplateApiUpdateTemplateProjectOpts{
-				Body:           optional.NewInterface(template),
-				HarnessAccount: optional.NewString(c.AccountId),
-			})
-			if resp.Identifier != "" {
-				template_id = resp.Identifier
+
+		if template_yaml != "" {
+			if project_id != "" {
+				resp, httpResp, err = c.ProjectTemplateApi.UpdateTemplateProject(ctx, project_id, id, org_id, version, &nextgen.ProjectTemplateApiUpdateTemplateProjectOpts{
+					Body:           optional.NewInterface(template),
+					HarnessAccount: optional.NewString(c.AccountId),
+				})
+				if resp.Identifier != "" {
+					template_id = resp.Identifier
+				} else {
+					template_id = resp.Slug
+				}
+			} else if org_id != "" && project_id == "" {
+				resp, httpResp, err = c.OrgTemplateApi.UpdateTemplateOrg(ctx, id, org_id, version, &nextgen.OrgTemplateApiUpdateTemplateOrgOpts{
+					HarnessAccount: optional.NewString(c.AccountId),
+					Body:           optional.NewInterface(template),
+				})
+				if resp.Identifier != "" {
+					template_id = resp.Identifier
+				} else {
+					template_id = resp.Slug
+				}
 			} else {
-				template_id = resp.Slug
+				resp, httpResp, err = c.AccountTemplateApi.UpdateTemplateAcc(ctx, id, version, &nextgen.AccountTemplateApiUpdateTemplateAccOpts{
+					HarnessAccount: optional.NewString(c.AccountId),
+					Body:           optional.NewInterface(template),
+				})
+				if resp.Identifier != "" {
+					template_id = resp.Identifier
+				} else {
+					template_id = resp.Slug
+				}
 			}
-		} else if org_id != "" && project_id == "" {
-			resp, httpResp, err = c.OrgTemplateApi.UpdateTemplateOrg(ctx, id, org_id, version, &nextgen.OrgTemplateApiUpdateTemplateOrgOpts{
-				HarnessAccount: optional.NewString(c.AccountId),
-				Body:           optional.NewInterface(template),
-			})
-			if resp.Identifier != "" {
-				template_id = resp.Identifier
+		} 
+		
+		if is_stable == true {
+			if project_id != "" {
+				_ , httpResp, err = c.ProjectTemplateApi.UpdateTemplateStableProject(ctx, project_id, id, org_id, version, &nextgen.ProjectTemplateApiUpdateTemplateStableProjectOpts{
+					Body:           optional.NewInterface(template),
+					HarnessAccount: optional.NewString(c.AccountId),
+				})
+			} else if org_id != "" && project_id == "" {
+				_ , httpResp, err = c.OrgTemplateApi.UpdateTemplateStableOrg(ctx, id, org_id, version, &nextgen.OrgTemplateApiUpdateTemplateStableOrgOpts{
+					HarnessAccount: optional.NewString(c.AccountId),
+					Body:           optional.NewInterface(template),
+				})
 			} else {
-				template_id = resp.Slug
+				_ , httpResp, err = c.AccountTemplateApi.UpdateTemplateStableAcc(ctx, id, version, &nextgen.AccountTemplateApiUpdateTemplateStableAccOpts{
+					HarnessAccount: optional.NewString(c.AccountId),
+					Body:           optional.NewInterface(template),
+				})
 			}
-		} else {
-			resp, httpResp, err = c.AccountTemplateApi.UpdateTemplateAcc(ctx, id, version, &nextgen.AccountTemplateApiUpdateTemplateAccOpts{
-				HarnessAccount: optional.NewString(c.AccountId),
-				Body:           optional.NewInterface(template),
-			})
-			if resp.Identifier != "" {
-				template_id = resp.Identifier
-			} else {
-				template_id = resp.Slug
-			}
+			template_id = id
 		}
 	}
 
@@ -375,6 +493,50 @@ func resourceTemplateCreateOrUpdate(ctx context.Context, d *schema.ResourceData,
 	readTemplate(d, respGet, comments, store_type, base_branch, commit_message, connector_ref)
 
 	return nil
+}
+
+func createImportFromGitRequestForTemplates(d *schema.ResourceData) *nextgen.TemplatesImportRequestBody {
+
+	template_git_import_details := &nextgen.GitImportDetails{}
+	if attr, ok := d.GetOk("git_import_details"); ok {
+		config := attr.([]interface{})[0].(map[string]interface{})
+		if attr, ok := config["branch_name"]; ok {
+			template_git_import_details.BranchName = attr.(string)
+		}
+		if attr, ok := config["file_path"]; ok {
+			template_git_import_details.FilePath = attr.(string)
+		}
+		if attr, ok := config["connector_ref"]; ok {
+			template_git_import_details.ConnectorRef = attr.(string)
+		}
+		if attr, ok := config["repo_name"]; ok {
+			template_git_import_details.RepoName = attr.(string)
+		}
+		if attr, ok := config["is_force_import"]; ok {
+			template_git_import_details.IsForceImport = attr.(bool)
+		}
+	}
+
+	template_import_request := &nextgen.TemplatesImportRequestDto{}
+	if attr, ok := d.GetOk("template_import_request"); ok {
+		config := attr.([]interface{})[0].(map[string]interface{})
+		if attr, ok := config["template_name"]; ok {
+			template_import_request.TemplateName = attr.(string)
+		}
+		if attr, ok := config["template_version"]; ok {
+			template_import_request.TemplateVersion = attr.(string)
+		}
+		if attr, ok := config["template_description"]; ok {
+			template_import_request.TemplateDescription = attr.(string)
+		}
+		
+	}
+
+	template_import_request_body := &nextgen.TemplatesImportRequestBody{}
+	template_import_request_body.GitImportDetails = template_git_import_details
+	template_import_request_body.TemplatesImportRequest = template_import_request
+
+	return template_import_request_body
 }
 
 func resourceTemplateDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {